ETA108 Development 3..

+++ trunk/WINCE600/PLATFORM/EM9170/SRC/DRIVERS/ETA108/ETA108Class.h
@@ -95,7 +95,7 @@ private:
 	HANDLE m_hADCEvent;		// ADC completed event
 	HANDLE m_hCSPIEvent;	// SPI transfer completed event
 
-	long  m_dwRxBufSeek;
+	DWORD  m_dwRxBufSeek;
 	DWORD  m_dwSamplingLength;
 	ADS8201_CFG m_stADS8201CFG;
 	BOOL   m_bTerminate;	// ADC moniter thread exit event
